requirements.txt
groq==0.25.0
python-dotenv==1.1.0
pydantic<3,>=1.9.0 